cylinder

+++ backend/controller/cylinderdepositController.js
@@ -2,32 +2,27 @@ const cylinderdepositService = require('../services/cylinderdepositService');
 const asyncHandler = require('../utils/asyncHandler');
 const ApiError =require('../utils/ApiError');
 
-exports.CreateBulkDeposit =asyncHandler(async(req,res,next)=>{
-    const cylinderdepositdata = req.body
-
-    Object.keys(cylinderdepositdata).forEach(key => {
-        if(cylinderdepositdata[key] === undefined || cylinderdepositdata[key] === null){
-            delete cylinderdepositdata[key];
-        }
-    });
-
-    const result = await cylinderdepositService.CreateBulkDeposit(cylinderdepositdata)
-    return res.status(201).json(result)
-})
-
 exports.findAllCylinderdeposit =asyncHandler(async(req,res,next)=>{
-    const result=await cylinderdepositService.findAllCylinderdepositd();
+    const result=await cylinderdepositService.findAllCylinderdeposit();
     res.status(200).json(result);
 })
 exports.updateCylinderdeposit = asyncHandler(async(req,res,next)=>{
-    const orderdata=req.body
 
-    const result =await cylinderdepositService.updateCylinderdeposit(orderdata);
+    const deposit_id = req.params.id
+    const {qty_return} = req.body
+    
+    if(!deposit_id || !qty_return) { return next(new ApiError(400, "deposit_id or qty_return are requried"))}
+
+    const result =await cylinderdepositService.updateCylinderdeposit(deposit_id , qty_return);
     res.status(200).json(result);
 })
 
-exports.deleteCylinderdeposit =asyncHandler(async(req,res,next)=>{
+exports.deleteCylinderdeposit = asyncHandler(async(req,res,next)=>{
+
     const {id}=req.params
+
+    if(!id) {return next(new ApiError(400, "id is required"))}
+
     const result =await cylinderdepositService.deleteCylinderdeposit(id);
     res.status(200).json(result);
 })
